requirements.txt
# Common libraries
loguru
mkdocs
pip
pytest
python-dotenv
pandas
requests
ruff
tqdm
typer
# Datasets
h2 
orjson 
pyarrow
